refactor: Move collectors from DebugBar/Collectors to Collector directory
- Move all collector files from src/DebugBar/Collectors/ to src/Collector/
- Update namespaces from LaminasMicroscope\DebugBar\Collectors to LaminasMicroscope\Collector
- Update all import statements in source files
- Update configuration files with new collector paths
- Update test files with new collector namespaces
- Update documentation references in CLAUDE.md
- Fix coding standards violations

This change consolidates all collectors in a single directory, making the
codebase structure cleaner and more intuitive. The DebugBar directory now
contains only DebugBar-specific code while collectors are organized separately.
