+++ src/Collector/PDOCollector.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace LaminasMicroscope\DebugBar\Collectors;
+namespace LaminasMicroscope\Collector;
 
 use DebugBar\DataCollector\AssetProvider;
 use DebugBar\DataCollector\DataCollector;
