+++ src/Controller/DashboardController.php
@@ -9,8 +9,8 @@ use Laminas\Http\Response;
 use Laminas\Mvc\Controller\AbstractActionController;
 use Laminas\View\Model\ViewModel;
 use LaminasMicroscope\Cache\CacheManager;
+use LaminasMicroscope\Collector\EnhancedPDOCollector;
 use LaminasMicroscope\Config\ConfigurationService;
-use LaminasMicroscope\DebugBar\Collectors\EnhancedPDOCollector;
 use LaminasMicroscope\Manager\ComponentManager;
 use LaminasMicroscope\Microscope\Storage\ReportStorage;
 use RuntimeException;
