+++ src/Factory/Controller/DashboardControllerFactory.php
@@ -7,9 +7,9 @@ namespace LaminasMicroscope\Factory\Controller;
 use Exception;
 use Laminas\ServiceManager\Factory\FactoryInterface;
 use LaminasMicroscope\Cache\CacheManager;
+use LaminasMicroscope\Collector\EnhancedPDOCollector;
 use LaminasMicroscope\Config\ConfigurationService;
 use LaminasMicroscope\Controller\DashboardController;
-use LaminasMicroscope\DebugBar\Collectors\EnhancedPDOCollector;
 use LaminasMicroscope\Manager\ComponentManager;
 use LaminasMicroscope\Microscope\Storage\ReportStorage;
 use Psr\Container\ContainerInterface;
