+++ tests/Integration/Controller/Phase3ControllerTest.php
@@ -10,7 +10,7 @@ use LaminasMicroscope\Controller\DashboardController;
 use LaminasMicroscope\Config\ConfigurationService;
 use LaminasMicroscope\Cache\CacheManager;
 use LaminasMicroscope\Microscope\Storage\ReportStorage;
-use LaminasMicroscope\DebugBar\Collectors\EnhancedPDOCollector;
+use LaminasMicroscope\Collector\EnhancedPDOCollector;
 
 class Phase3ControllerTest extends AbstractHttpControllerTestCase
 {
