+++ tests/bootstrap.php
@@ -87,19 +87,19 @@ class TestHelper
                         'factory' => 'direct',
                     ],
                     'config'   => [
-                        'class'        => \LaminasMicroscope\DebugBar\Collectors\LaminasConfigCollector::class,
+                        'class'        => \LaminasMicroscope\Collector\LaminasConfigCollector::class,
                         'factory'      => 'service',
-                        'service_name' => \LaminasMicroscope\DebugBar\Collectors\LaminasConfigCollector::class,
+                        'service_name' => \LaminasMicroscope\Collector\LaminasConfigCollector::class,
                     ],
                     'pdo'      => [
-                        'class'        => \LaminasMicroscope\DebugBar\Collectors\PDOCollector::class,
+                        'class'        => \LaminasMicroscope\Collector\PDOCollector::class,
                         'factory'      => 'service',
-                        'service_name' => \LaminasMicroscope\DebugBar\Collectors\PDOCollector::class,
+                        'service_name' => \LaminasMicroscope\Collector\PDOCollector::class,
                     ],
                     'request'  => [
-                        'class'        => \LaminasMicroscope\DebugBar\Collectors\LaminasRequestCollector::class,
+                        'class'        => \LaminasMicroscope\Collector\LaminasRequestCollector::class,
                         'factory'      => 'service',
-                        'service_name' => \LaminasMicroscope\DebugBar\Collectors\LaminasRequestCollector::class,
+                        'service_name' => \LaminasMicroscope\Collector\LaminasRequestCollector::class,
                     ],
                 ],
             ],
