+++ tests/config/testing.global.php
@@ -64,19 +64,19 @@ return [
                 'factory' => 'direct',
             ],
             'config'   => [
-                'class'        => \LaminasMicroscope\DebugBar\Collectors\LaminasConfigCollector::class,
+                'class'        => \LaminasMicroscope\Collector\LaminasConfigCollector::class,
                 'factory'      => 'service',
-                'service_name' => \LaminasMicroscope\DebugBar\Collectors\LaminasConfigCollector::class,
+                'service_name' => \LaminasMicroscope\Collector\LaminasConfigCollector::class,
             ],
             'pdo'      => [
-                'class'        => \LaminasMicroscope\DebugBar\Collectors\PDOCollector::class,
+                'class'        => \LaminasMicroscope\Collector\PDOCollector::class,
                 'factory'      => 'service',
-                'service_name' => \LaminasMicroscope\DebugBar\Collectors\PDOCollector::class,
+                'service_name' => \LaminasMicroscope\Collector\PDOCollector::class,
             ],
             'request'  => [
-                'class'        => \LaminasMicroscope\DebugBar\Collectors\LaminasRequestCollector::class,
+                'class'        => \LaminasMicroscope\Collector\LaminasRequestCollector::class,
                 'factory'      => 'service',
-                'service_name' => \LaminasMicroscope\DebugBar\Collectors\LaminasRequestCollector::class,
+                'service_name' => \LaminasMicroscope\Collector\LaminasRequestCollector::class,
             ],
         ],
     ],
